requirements.txt
fastapi==0.100.1
uvicorn[standard]==0.29.0
sqlmodel==0.0.16
python-dotenv == 1.0.1
psycopg2
sqlalchemy
